Connect GeoGirafe & Georama (#91)
This PR allows a user to directly publish a GeoGirafe Theme from an
already integrated QGIS Project. Original tree of the QGIS Project is
kept and all layers are added into their corresponding groups. The main
changes are:
- add shortcut button to publish qgis project as geogirafe theme
- make a dedicated endpoint to publish webgis WMS (`/webgis/maps`) this
includes separate permission handling and complete separation of the
WMS/OAPIF enpoint publishing and the WebGIS publishing
- Use as much as possible implementaion from `maps` app through
subclassing to avoid reimplementation
- fix endpoint of `themes.json`
- offer a first simple translation endpoint
`/webgis/translations/de.json` which dynamically translates the machine
`name` attributes with the `title` attribute of `PublishedAsTheme`,
`LayerGroupMp` and `PublishedAsLayerWms` currently.