Set GEORAMA_DEBUG=true in .env.example (#113)
The `.env.example` currently serves as an example for local development,
not for production (an example for production doesn't really exist yet
as such, as far as I am aware - see #107).

Therefore enabling `GEORAMA_DEBUG` by default in that scenario is more
consistent.

This fixes an issue where, if you just follow the current README, the
static assets wouldn't be served by the dev server, and they 404,
because `GEORAMA_DEBUG` was turned off in the example env.