Move settings into django-configurations base class:
This just idents the existing settings one level to the right, making them
part of the `Base` configuration class. No changes are being made yet to how
the settings are read from the environment variables, how they're typed, and
what defaults they have. That will be done in the next commits.