Move CSRF_TRUSTED_ORIGINS from .env to settings.py:
This moves the definition of CSRF trusted origins (for the Dev setup) out of
the .env.example and into the `Dev` config in settings.py

Again, I believe many of these origins might not be needed anymore, and this
list is not actually correct as of now, but in this PR we'll just move the
settings over, and preserve the current behavior.

As part of #108 we'll look at CORS and CSRF origins and clean them up.