Clean out unnecessary CORS / CSRF trusted origins:
This removes some whitelisted CORS and CSRF origins that are almost certainly
not necessary (any more).

http://localhost:4242 is the default backend URL, so this doesn't need to be
in CSRF_TRUSTED_ORIGINS - only *cross* origin requests need to be whitelisted.

http[s]://localhost seems like an origin that would only be used in a setup
with a reverse proxy in front, and smells of a production or staging system.

For those cases, those should not be put in the `Dev` configuration, but in
a more appropriate config instead.