Load env vars from .env using python-dotenv:
Extend manage.py to load environment variables from an .env file, if present.

This is relevant when developing georama directly on the local filesystem,
instead of in a Docker container:

When using Docker to develop, docker compose will load the .env file, and
pass the environment variables on to Django. When not using Docker for
development however, that doesn't happen automatically, because docker compose
is not in play (for Django at least).

Therefore, in order to have a local development setup behave the same as a
docker based development setup, given the same .env, we also need to load
env vars from the .env file when using Django directly.

This fixes the local development workflow when using `make serve-dev`, but
also when using commands like `make make-migrations`.