Refactor OGC API Features to a class based approach (#140)
The API Endpoint for OGC API Features through pygeoapi is a generic way
we want to reuse throughout the Georama eco system. Currently mostly in
the webgis app part.

Therefore we turn it into a class based view. This way we can create
subclasses of it to reuse it as a snippet easily.

fixes #120