Fix template loading:
Because both APP_DIRS was enabled as well as a manual include for
georama.core's template directory, our templates were processed twice.

This caused an issue with template overrides that inherit from the template
they override, and use {{ block.super }} to selectively add content:

Because our templates were processed twice, the addition of the content was
also processed twice, leading to all added content being duplicated.

If APP_DIRS is set to True (which it should, in almost all cases), the
directories added to DIRS must NOT be *app* template directories, but only
*project* template directories outside of apps. Otherwise they get processed
twice.