Make SOCIALACCOUNT_PROVIDERS settings property:
This makes the SOCIALACCOUNT_PROVIDERS setting a property that gets its
details from individual settings, which then can also be controlled via
env vars individually.

This makes it much easier to for example just change the OIDC server URL
by overriding a single setting, instead of having to override a dictionary as
a multiline string in an env var.